backend/pyproject.toml
[tool.poetry]
name = "dependencias do projeto"
version = "0.1.0"
description = "Bibliotecas e Frameworks necessários para execução do projeto"
authors = ["RuanCarlosSilva <143547755+ruansilva123@users.noreply.github.com>"]
license = "MIT Licence"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.2"
pdfplumber = "^0.11.4"
google-generativeai = "^0.8.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
